# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenXLR Local Control API",
    "version": "1.0.0",
    "description": "Version 1 of the loopback-only OpenXLR daemon API. The daemon is intentionally bound to 127.0.0.1 and has no browser CORS policy or remote-network authentication."
  },
  "servers": [
    { "url": "http://127.0.0.1:37890", "description": "Default local daemon" }
  ],
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Liveness check",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "The daemon HTTP loop is responsive.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Health" } } }
          }
        }
      }
    },
    "/api/v1": {
      "get": {
        "summary": "Discover API v1 resources",
        "operationId": "getApiIndex",
        "responses": {
          "200": {
            "description": "Stable resource URLs and daemon version.",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/api/v1/state": {
      "get": {
        "summary": "Read the authoritative device and mixer state",
        "operationId": "getState",
        "responses": {
          "200": {
            "description": "The same state shape sent by the event WebSocket.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StateEvent" } } }
          }
        }
      }
    },
    "/api/v1/plugins": {
      "get": {
        "summary": "List isolated LV2 and native Linux VST3 plug-ins",
        "description": "Returns cached ready, failed, rejected, and quarantined scan results. Rescans run in bounded helper processes.",
        "operationId": "listPlugins",
        "responses": {
          "200": {
            "description": "Installed effects and their control ports.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PluginsEvent" } } }
          }
        }
      }
    },
    "/api/v1/presets": {
      "get": {
        "summary": "List reusable effect-chain and individual plug-in presets",
        "operationId": "listPresets",
        "responses": { "200": { "description": "Preset names grouped by kind.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PresetsEvent" } } } } }
      }
    },
    "/api/v1/presets/{kind}/{name}": {
      "get": {
        "summary": "Export a versioned preset JSON document",
        "operationId": "exportPreset",
        "parameters": [
          { "name": "kind", "in": "path", "required": true, "schema": { "type": "string", "enum": ["chains", "plugins"] } },
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string", "maxLength": 80 } }
        ],
        "responses": { "200": { "description": "Portable bounded preset JSON." }, "404": { "description": "Invalid kind/name or missing preset." } }
      }
    },
    "/api/v1/presets/{kind}/import": {
      "post": {
        "summary": "Validate and import an untrusted preset document",
        "operationId": "importPreset",
        "parameters": [
          { "name": "kind", "in": "path", "required": true, "schema": { "type": "string", "enum": ["chains", "plugins"] } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "201": { "description": "Preset validated and stored." }, "422": { "description": "Malformed, oversized, unsafe, or unsupported preset." } }
      }
    },
    "/api/v1/commands": {
      "post": {
        "summary": "Execute one hardware, mixer, routing, effect, or profile command",
        "description": "The response always has a requestId. Mutations also return the authoritative resulting state. A 422 response is a valid command that OpenXLR rejected; no retry should occur until its error is handled.",
        "operationId": "executeCommand",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Command" },
              "examples": {
                "setSendLevel": { "value": { "cmd": "setLevel", "channel": "music", "mix": "stream", "value": 0.72, "requestId": "my-app-42" } },
                "routeApplication": { "value": { "cmd": "assignApp", "identity": "org.mozilla.firefox", "channel": "browser" } },
                "addOutputMix": { "value": { "cmd": "createMix", "name": "Podcast" } },
                "routeMix": { "value": { "cmd": "setOutputRoute", "mix": "stream", "destinationId": "output:stable-id", "enabled": true, "stage": "MixProcessed" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command succeeded.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommandResult" } } }
          },
          "400": {
            "description": "Malformed JSON, a non-object request, or an invalid requestId.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEvent" } } }
          },
          "413": {
            "description": "The command exceeded 65536 bytes.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEvent" } } }
          },
          "422": {
            "description": "The command was parsed but rejected by validation, hardware, or the audio graph.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommandResult" } } }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Health": {
        "type": "object",
        "required": ["status", "apiVersion", "daemonVersion"],
        "properties": {
          "status": { "const": "ok" },
          "apiVersion": { "const": "1" },
          "daemonVersion": { "type": "string" }
        }
      },
      "Command": {
        "type": "object",
        "required": ["cmd"],
        "additionalProperties": true,
        "properties": {
          "cmd": {
            "type": "string",
            "description": "Command discriminator. Required and case-sensitive.",
            "enum": [
              "getState", "getDiagnostics", "listPlugins", "listPresets", "rescanPlugins",
              "retryPlugin", "unquarantinePlugin", "set", "setLevel",
              "setChannelMuted", "setMixVolume", "setMixMuted", "createChannel",
              "renameChannel", "deleteChannel", "reorderChannels", "createMix",
              "renameMix", "deleteMix", "reorderMixes", "assignStream", "assignApp",
              "forgetApp", "setMonitorOutput", "setMonitorOutputs", "setOutputRoute", "setMonitoredMix",
              "setOutputVolume", "setEnforcedDefaults", "setAuxPortEnabled", "setLowCutHz",
              "setSoftClipGuard", "setInserts", "showInsertUi", "setInsertBypass",
              "setInsertParam", "retryInsertHost", "unquarantineInsert",
              "saveChainPreset", "loadChainPreset", "savePluginPreset", "loadPluginPreset",
              "renamePreset", "duplicatePreset", "deletePreset",
              "setActiveDevice", "saveProfile", "loadProfile", "deleteProfile"
            ]
          },
          "requestId": { "type": "string", "maxLength": 128, "description": "Caller-generated correlation ID. HTTP generates one when omitted; WebSocket echoes it in commandResult." },
          "control": { "type": "string", "description": "Hardware control name for set. See docs/api.md for the supported names." },
          "value": { "description": "Boolean or numeric command value. Levels are linear 0..1 unless a command explicitly uses dB or Hz.", "oneOf": [{ "type": "boolean" }, { "type": "number" }, { "type": "null" }] },
          "channel": { "type": "string", "description": "Stable channel ID, or mix:<id> for an output insert chain." },
          "mix": { "type": "string", "description": "Stable output mix ID." },
          "streamId": { "type": "integer", "minimum": 0, "description": "Live PipeWire stream ID for assignStream." },
          "device": { "type": ["string", "null"], "description": "PipeWire node.name for a monitor output, or USB vvvv:pppp ID for setActiveDevice." },
          "devices": { "type": "array", "uniqueItems": true, "items": { "type": "string" }, "description": "Complete selected monitor-output node list." },
          "destinationId": { "type": "string", "description": "Stable output ID from mixer.routingDestinations; never substitute a display name or PipeWire object serial." },
          "enabled": { "type": "boolean", "description": "Whether the setOutputRoute matrix cell should exist." },
          "stage": { "type": "string", "enum": ["Raw", "HardwareFx", "PluginFx", "FullFx", "MixProcessed"], "description": "Processing tap for setOutputRoute. Use only a value advertised by the destination/cell capabilities." },
          "order": { "type": "array", "uniqueItems": true, "items": { "type": "string" }, "description": "Complete ordered set of editable channel or mix IDs." },
          "identity": { "type": "string", "description": "Stable application identity for assignApp or forgetApp." },
          "label": { "type": "string", "description": "Optional application or insert display label." },
          "sink": { "type": ["string", "null"], "description": "System default sink node to enforce; null disables sink enforcement." },
          "source": { "type": ["string", "null"], "description": "System default source node to enforce; null disables source enforcement." },
          "name": { "type": "string", "description": "Display name for create/rename or profile name for save/load/delete." },
          "newName": { "type": "string", "description": "Destination name for preset rename or duplicate." },
          "presetKind": { "type": "string", "enum": ["chain", "plugin"], "description": "Reusable preset family." },
          "inserts": { "type": "array", "items": { "$ref": "#/components/schemas/Insert" }, "description": "The complete ordered insert chain." },
          "insertId": { "type": "string", "description": "Stable ID of an insert in the selected chain." },
          "symbol": { "type": "string", "description": "LV2 control-port symbol for setInsertParam." }
        }
      },
      "Insert": {
        "type": "object",
        "required": ["id", "kind", "plugin"],
        "additionalProperties": true,
        "properties": {
          "id": { "type": "string" },
          "kind": { "type": "string", "enum": ["lv2", "vst3"] },
          "plugin": { "type": "string", "description": "Stable format-native plugin identity." },
          "label": { "type": "string" },
          "bypass": { "type": "boolean", "default": false },
          "params": { "type": "object", "additionalProperties": { "type": "number" } },
          "state": { "type": ["string", "null"], "contentEncoding": "base64", "description": "Bounded opaque format-native state." },
          "sidechains": { "type": "object", "additionalProperties": { "type": "string" }, "description": "Aux bus ID to stable OpenXLR signal source ID." }
        }
      },
      "CommandResult": {
        "type": "object",
        "required": ["type", "apiVersion", "requestId", "ok"],
        "properties": {
          "type": { "const": "commandResult" },
          "apiVersion": { "const": "1" },
          "requestId": { "type": "string" },
          "ok": { "type": "boolean" },
          "error": { "type": "string" },
          "result": { "description": "Typed query result for getState, getDiagnostics, or listPlugins." },
          "state": { "$ref": "#/components/schemas/StateEvent", "description": "Authoritative post-command state, present for HTTP mutations." }
        }
      },
      "StateEvent": {
        "type": "object",
        "required": ["type", "apiVersion", "daemonVersion", "features", "connected"],
        "properties": {
          "type": { "const": "state" },
          "apiVersion": { "const": "1" },
          "daemonVersion": { "type": "string" },
          "features": { "type": "array", "items": { "type": "string" } },
          "connected": { "type": "boolean" },
          "device": { "type": ["object", "null"] },
          "capabilities": { "type": ["object", "null"] },
          "state": { "type": ["object", "null"], "description": "Hardware controls for the active interface." },
          "mixer": { "type": ["object", "null"], "description": "Channels, mixes, sends, mutes, applications, routing, monitor selection, and inserts." },
          "devices": { "type": ["array", "null"], "items": { "type": "object" } },
          "profiles": { "type": ["array", "null"], "items": { "type": "string" } },
          "activeProfile": { "type": ["string", "null"] },
          "detected": { "type": ["array", "null"], "items": { "type": "object" } }
        },
        "additionalProperties": true
      },
      "MetersEvent": {
        "type": "object",
        "required": ["type", "apiVersion", "levels"],
        "properties": {
          "type": { "const": "meters" },
          "apiVersion": { "const": "1" },
          "levels": {
            "type": "object",
            "description": "Keys are ch:<channelId> or mix:<mixId>; values are stereo peaks in 0..1.",
            "additionalProperties": { "type": "array", "minItems": 2, "maxItems": 2, "items": { "type": "number", "minimum": 0, "maximum": 1 } }
          }
        }
      },
      "PluginsEvent": {
        "type": "object",
        "required": ["type", "apiVersion", "plugins"],
        "properties": {
          "type": { "const": "plugins" },
          "apiVersion": { "const": "1" },
          "plugins": { "type": "array", "items": { "type": "object" } }
        }
      },
      "PresetsEvent": {
        "type": "object",
        "required": ["type", "apiVersion", "chains", "plugins"],
        "properties": {
          "type": { "const": "presets" },
          "apiVersion": { "const": "1" },
          "chains": { "type": "array", "items": { "type": "string" } },
          "plugins": { "type": "array", "items": { "type": "string" } }
        }
      },
      "ErrorEvent": {
        "type": "object",
        "required": ["type", "apiVersion", "message"],
        "properties": {
          "type": { "const": "error" },
          "apiVersion": { "const": "1" },
          "message": { "type": "string" }
        }
      }
    }
  },
  "x-websocket": {
    "url": "ws://127.0.0.1:37890/api/v1/events",
    "legacyAlias": "ws://127.0.0.1:37890/ws",
    "description": "The first server frame is StateEvent. Later frames are StateEvent, MetersEvent, PluginsEvent, PresetsEvent, CommandResult, or ErrorEvent. Clients send Command objects. A correlated mutation receives authoritative state before commandResult.",
    "serverMessages": {
      "oneOf": [
        { "$ref": "#/components/schemas/StateEvent" },
        { "$ref": "#/components/schemas/MetersEvent" },
        { "$ref": "#/components/schemas/PluginsEvent" },
        { "$ref": "#/components/schemas/PresetsEvent" },
        { "$ref": "#/components/schemas/CommandResult" },
        { "$ref": "#/components/schemas/ErrorEvent" }
      ]
    },
    "clientMessages": { "$ref": "#/components/schemas/Command" }
  }
}